ft/implement-logging
