+++ src/Hahn.ApplicatonProcess.December2020.Domain/Entities/BaseEntity.cs
@@ -1,11 +1,5 @@
 namespace Hahn.ApplicatonProcess.December2020.Domain.Entities
 {
-    using System;
-    using System.Collections.Generic;
-    using System.Linq;
-    using System.Text;
-    using System.Threading.Tasks;
-
     public abstract class BaseEntity
     {
         public virtual int Id { get; set; }
