+++ src/Hahn.ApplicatonProcess.December2020.Web/Controllers/BaseApiController.cs
@@ -1,14 +1,11 @@
+using Microsoft.AspNetCore.Mvc;
+
 namespace Hahn.ApplicatonProcess.December2020.Web.Controllers
 {
-    using System;
-    using System.Collections.Generic;
-    using System.Linq;
-    using System.Threading.Tasks;
-    using Microsoft.AspNetCore.Mvc;
-
     [ApiController]
     [Route("[controller]")]
     public class BaseApiController : ControllerBase
-    { 
+    {
+
     }
 }
