+++ src/Hahn.ApplicatonProcess.December2020.Web/appsettings.json
@@ -9,5 +9,19 @@
       "Microsoft.Hosting.Lifetime": "Information"
     }
   },
+  "Serilog": {
+    "Using": [ "Serilog.Sinks.Console", "Serilog.Sinks.File" ],
+    "WriteTo": [
+      { "Name": "Console" },
+      {
+        "Name": "File",
+        "Args": { "path": "Logs/log.txt" }
+      }
+    ],
+    "Enrich": [ "FromLogContext", "WithMachineName", "WithThreadId" ],
+    "Properties": {
+      "Application": "Sample"
+    }
+  },
   "AllowedHosts": "*"
 }
